댓글을 2페이지 넘게 작성했을때, 1페이지에 댓글을 추가하면 2페이지로 넘어가서 보여주는 현상 고치기
