+++ dokky/src/main/java/org/my/mapper/BoardMapper.java
@@ -82,6 +82,4 @@ public interface BoardMapper {
 	
 	public void updateReplyCnt(@Param("board_num") Long board_num, @Param("amount") int amount);
 
-	public int[] getReplyNums(Long board_num);
-	
 }
