+++ dokky/src/main/resources/org/my/mapper/BoardMapper.xml
@@ -283,10 +283,6 @@
 		select * from DK_BOARD where board_num = #{board_num}
 	</select>
 	
-	<select id="getReplyNums" resultType="int">
-		select reply_num from dk_reply where board_num = #{reply_num} order by group_num ASC, order_step ASC
-	</select>
-	
 	<update id="updateBoard">
 		update DK_BOARD 
 			set title= #{title},
